ios: fix timezone
The same issue we fixed on Android also exists on iOS. See comments in
code for details.
