+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Fix potential crash in the AOPP workflow
 - Windows: fix send/feetarget dropdown UI rendering bug
 - More efficient account initialization by fetching all account Bitcoin xpubs at once
+- iOS: fix wrong timezone when confirming time on BitBox02 (it would always show the time in UTC)
 
 ## v4.48.4
 - macOS: fix potential USB communication issue with BitBox02 bootloaders <v1.1.2 and firmwares <v9.23.1
