+++ backend/mobileserver/fixtimezone_android.go
@@ -1,0 +1,42 @@
+// Copyright 2025 Shift Crypto AG
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build android
+
+package mobileserver
+
+import (
+	"os/exec"
+	"strings"
+	"time"
+)
+
+// fixTimezone sets the local timezone on Android. This is a workaround to the bug that on Android,
+// time.Local is hard-coded to UTC. See https://github.com/golang/go/issues/20455.
+//
+// We need the correct timezone to be able to send the `time.Now().Zone()` offset to the BitBox02.
+// Without it, the BitBox02 will always display UTC time instad of local time.
+//
+// This fix is copied from https://github.com/golang/go/issues/20455#issuecomment-342287698.
+func fixTimezone() {
+	out, err := exec.Command("/system/bin/getprop", "persist.sys.timezone").Output()
+	if err != nil {
+		return
+	}
+	z, err := time.LoadLocation(strings.TrimSpace(string(out)))
+	if err != nil {
+		return
+	}
+	time.Local = z
+}
